Refactor SafeListSidebar uses redux-hooks approach
